removes 'rev' and 'ref' attributes in meta
With the move to flakes we now encourage are
encouraged to use pure evaluation of nix exprressions
and the current meta machinery was using the impure
state of the filesystem to determine which git
ref/rev was being used. We'll have to find a way to
thread this information down from the flake root.